Complete multi-algorithm testing.
Complete multi-algorithm testing. Wall-follower is the worst performing
algorithm with no optimization. Basic waterfall requires a large
exploration phase to take full advantage of the speed phase. Search
waterfall immediately stops exploring when an ideal route is located,
Oracle waterfall is used as an ideal sample, using foreknowledge of the
maze to accomplish a best case score. Built into the robot class in
robot.py is a random choice algorithm that fails to find the center in
size 14 and larger mazes within the time limit.